kl koodattu

+++ Tammipeli/src/main/java/jalotammi/tammipeli/domain/Lukupari.java
@@ -3,18 +3,18 @@
 package jalotammi.tammipeli.domain;
 
 public class Lukupari {
-    public double x, y;
+    public int x, y;
 
-    public Lukupari(double x, double y) {
+    public Lukupari(int x, int y) {
       this.x = x;
       this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
